# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MinePilot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /servers:
    get:
      operationId: listServers
      tags: [servers]
      summary: List user's servers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Server"
    post:
      operationId: createServer
      tags: [servers]
      summary: Add a server
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServerRequest"
      responses:
        "201":
          description: Server created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
  /servers/{id}:
    get:
      operationId: getServer
      tags: [servers]
      summary: Get server by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    put:
      operationId: updateServer
      tags: [servers]
      summary: Update server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServerRequest"
      responses:
        "200":
          description: Server updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Server"
    delete:
      operationId: deleteServer
      tags: [servers]
      summary: Delete server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/connect:
    post:
      operationId: connectServer
      tags: [servers]
      summary: Connect to server via SSH
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionStatus"
  /servers/{id}/disconnect:
    post:
      operationId: disconnectServer
      tags: [servers]
      summary: Disconnect from server
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /servers/{id}/stats:
    get:
      operationId: getServerStats
      tags: [servers]
      summary: Get server stats (CPU, RAM, TPS, players)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Server stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerStats"
  /servers/{id}/setup:
    post:
      operationId: setupServer
      tags: [servers]
      summary: Run beginner server setup wizard
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupServerRequest"
      responses:
        "200":
          description: Setup initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Server:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        host:
          type: string
        sshPort:
          type: integer
        sshUsername:
          type: string
        status:
          type: string
          enum: [online, offline, connecting, error]
        serverType:
          type: string
          nullable: true
        gameMode:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, host, sshPort, sshUsername, status, createdAt]
    CreateServerRequest:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        sshPort:
          type: integer
          default: 22
        sshUsername:
          type: string
        sshPassword:
          type: string
        serverType:
          type: string
          nullable: true
        gameMode:
          type: string
          nullable: true
      required: [name, host, sshPort, sshUsername, sshPassword]
    UpdateServerRequest:
      type: object
      properties:
        name:
          type: string
        host:
          type: string
        sshPort:
          type: integer
        sshUsername:
          type: string
        sshPassword:
          type: string
        serverType:
          type: string
          nullable: true
        gameMode:
          type: string
          nullable: true
    ConnectionStatus:
      type: object
      properties:
        connected:
          type: boolean
        message:
          type: string
      required: [connected, message]
    ServerStats:
      type: object
      properties:
        online:
          type: boolean
        playerCount:
          type: integer
        maxPlayers:
          type: integer
        ramUsed:
          type: number
        ramTotal:
          type: number
        cpuUsage:
          type: number
        tps:
          type: number
        diskUsed:
          type: number
        diskTotal:
          type: number
      required: [online, playerCount, maxPlayers, ramUsed, ramTotal, cpuUsage, tps, diskUsed, diskTotal]
    SetupServerRequest:
      type: object
      properties:
        serverType:
          type: string
          enum: [paper, spigot, purpur]
        gameMode:
          type: string
          enum: [survival, skyblock, smp]
        installPlugins:
          type: array
          items:
            type: string
      required: [serverType, gameMode]
...