# lib/api-zod/src/generated/types/server.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinePilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServerStatus } from "./serverStatus";

export interface Server {
  id: number;
  userId: number;
  name: string;
  host: string;
  sshPort: number;
  sshUsername: string;
  status: ServerStatus;
  serverType?: string | null;
  gameMode?: string | null;
  createdAt: Date;
}