# lib/api-zod/src/generated/types/serverStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinePilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServerStats {
  online: boolean;
  playerCount: number;
  maxPlayers: number;
  ramUsed: number;
  ramTotal: number;
  cpuUsage: number;
  tps: number;
  diskUsed: number;
  diskTotal: number;
}