# lib/api-zod/src/generated/types/setupServerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinePilot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SetupServerRequestGameMode } from "./setupServerRequestGameMode";
import type { SetupServerRequestServerType } from "./setupServerRequestServerType";

export interface SetupServerRequest {
  serverType: SetupServerRequestServerType;
  gameMode: SetupServerRequestGameMode;
  installPlugins?: string[];
}