# lib/api-zod/src/generated/types/updateServerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MinePilot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateServerRequest {
  name?: string;
  host?: string;
  sshPort?: number;
  sshUsername?: string;
  sshPassword?: string;
  serverType?: string | null;
  gameMode?: string | null;
}